# static/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "component API",
    "version": "1.0.0",
    "description": "API for managing Bills of Materials (components)"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
    "/component/{component_id}": {
      "put": {
        "tags": [
          "component"
        ],
        "summary": "Update a component",
        "description": "Updates an existing component in the database and in an Excel file.",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "description": "ID of the component to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "New values for the component",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "section": {
                    "type": "string",
                    "description": "The section of the component"
                  },
                  "sub-section": {
                    "type": "string",
                    "description": "The sub-section of the component"
                  },
                  "level": {
                    "type": "string",
                    "description": "The level of the component"
                  },
                  "TCode": {
                    "type": "string",
                    "description": "The T-code of the component"
                  },
                  "code": {
                    "type": "string",
                    "description": "The code of the component"
                  },
                  "Parent": {
                    "type": "string",
                    "description": "The parent code of the component"
                  },
                  "module": {
                    "type": "boolean",
                    "description": "Whether the component is a module or not"
                  },
                  "repeated": {
                    "type": "boolean",
                    "description": "Whether the component is repeated or not"
                  },
                  "condition": {
                    "type": "string",
                    "description": "The condition of the component"
                  },
                  "fig": {
                    "type": "string",
                    "description": "The figure of the component"
                  },
                  "position": {
                    "type": "string",
                    "description": "The position of the component"
                  },
                  "part": {
                    "type": "string",
                    "description": "The part number of the component"
                  },
                  "quantity": {
                    "type": "integer",
                    "description": "The quantity of the component"
                  },
                  "standard": {
                    "type": "string",
                    "description": "The standard of the component"
                  },
                  "national_stock_number": {
                    "type": "string",
                    "description": "The national stock number of the component"
                  },
                  "part_number": {
                    "type": "string",
                    "description": "The part number of the component"
                  },
                  "ATA_number": {
                    "type": "string",
                    "description": "The ATA number of the component"
                  },
                  "weight_on_pcs": {
                    "type": "number",
                    "description": "The weight/pcs of the component"
                  },
                  "weight": {
                    "type": "number",
                    "description": "The weight of the component"
                  },
                  "version": {
                    "type": "number",
                    "description": "The version of the component"
                  },
                  "sap_code": {
                    "type": "number",
                    "description": "The sap code of the component"
                  },
                  "head_of_CVE": {
                    "type": "number",
                    "description": "The head of cve of the component"
                  },
                  "CP": {
                    "type": "boolean",
                    "description": "The cp of the component"
                  },
                  "CCL": {
                    "type": "boolean",
                    "description": "The ccl of the component"
                  },
                  "drawing_lvl_2": {
                    "type": "boolean",
                    "description": "The drawing level 2 of the component"
                  },
                  "ITP_and_SPc_of_material": {
                    "type": "boolean",
                    "description": "The ITP and SPc of material of the component"
                  },
                  "ITP_and_SPC_of_production_technology": {
                    "type": "boolean",
                    "description": "The ITP and SPC of production technology of the component"
                  },
                  "type_of_analysis": {
                    "type": "boolean",
                    "description": "The type of analysis of the component"
                  },
                  "prototype_part": {
                    "type": "boolean",
                    "description": "The prototype of the component"
                  },
                  "test": {
                    "type": "boolean",
                    "description": "The test of the component"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "component"
        ],
        "summary": "Shows a component",
        "description": "Shows an existing component in the database using its id.",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "description": "ID of the component to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "New values for the component",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "section": {
                    "type": "string",
                    "description": "The section of the component"
                  },
                  "sub-section": {
                    "type": "string",
                    "description": "The sub-section of the component"
                  },
                  "level": {
                    "type": "string",
                    "description": "The level of the component"
                  },
                  "TCode": {
                    "type": "string",
                    "description": "The T-code of the component"
                  },
                  "code": {
                    "type": "string",
                    "description": "The code of the component"
                  },
                  "Parent": {
                    "type": "string",
                    "description": "The parent code of the component"
                  },
                  "module": {
                    "type": "boolean",
                    "description": "Whether the component is a module or not"
                  },
                  "repeated": {
                    "type": "boolean",
                    "description": "Whether the component is repeated or not"
                  },
                  "condition": {
                    "type": "string",
                    "description": "The condition of the component"
                  },
                  "fig": {
                    "type": "string",
                    "description": "The figure of the component"
                  },
                  "position": {
                    "type": "string",
                    "description": "The position of the component"
                  },
                  "part": {
                    "type": "string",
                    "description": "The part number of the component"
                  },
                  "quantity": {
                    "type": "integer",
                    "description": "The quantity of the component"
                  },
                  "standard": {
                    "type": "string",
                    "description": "The standard of the component"
                  },
                  "national_stock_number": {
                    "type": "string",
                    "description": "The national stock number of the component"
                  },
                  "part_number": {
                    "type": "string",
                    "description": "The part number of the component"
                  },
                  "ATA_number": {
                    "type": "string",
                    "description": "The ATA number of the component"
                  },
                  "weight_on_pcs": {
                    "type": "number",
                    "description": "The weight/pcs of the component"
                  },
                  "weight": {
                    "type": "number",
                    "description": "The weight of the component"
                  },
                  "version": {
                    "type": "number",
                    "description": "The version of the component"
                  },
                  "sap_code": {
                    "type": "number",
                    "description": "The sap code of the component"
                  },
                  "head_of_CVE": {
                    "type": "number",
                    "description": "The head of cve of the component"
                  },
                  "CP": {
                    "type": "boolean",
                    "description": "The cp of the component"
                  },
                  "CCL": {
                    "type": "boolean",
                    "description": "The ccl of the component"
                  },
                  "drawing_lvl_2": {
                    "type": "boolean",
                    "description": "The drawing level 2 of the component"
                  },
                  "ITP_and_SPc_of_material": {
                    "type": "boolean",
                    "description": "The ITP and SPc of material of the component"
                  },
                  "ITP_and_SPC_of_production_technology": {
                    "type": "boolean",
                    "description": "The ITP and SPC of production technology of the component"
                  },
                  "type_of_analysis": {
                    "type": "boolean",
                    "description": "The type of analysis of the component"
                  },
                  "prototype_part": {
                    "type": "boolean",
                    "description": "The prototype of the component"
                  },
                  "test": {
                    "type": "boolean",
                    "description": "The test of the component"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}